requirements.txt
ffmpeg-python
torch
torchaudio
torchvision
pyannote.audio
speechbrain
huggingface_hub
tk
